requirements.txt
hurry.filesize
dnspython
pymongo
pillow
flask